changes for building image on s390x

+++ images/ubuntu/scripts/build/install-cmake.sh
@@ -13,19 +13,24 @@ echo "Checking to see if the installer script has already been run"
 if command -v cmake; then
     echo "cmake is already installed"
 else
-	# Download script to install CMake
-	download_url=$(resolve_github_release_asset_url "Kitware/CMake" "endswith(\"inux-$ARCH_L.sh\")" "latest")
-	curl -fsSL "${download_url}" -o cmakeinstall.sh
+    if [[ "$(uname -m)" == "s390x" ]]; then
+        echo "Detected s390x architecture. Installing CMake via APT..."
+        apt-get update
+        apt-get install -y cmake
+    else
+        download_url=$(resolve_github_release_asset_url "Kitware/CMake" "endswith(\"inux-$ARCH_L.sh\")" "latest")
+        curl -fsSL "${download_url}" -o cmakeinstall.sh
 
-	# Supply chain security - CMake
-	hash_url=$(resolve_github_release_asset_url "Kitware/CMake" "endswith(\"SHA-256.txt\")" "latest")
-	external_hash=$(get_checksum_from_url "$hash_url" "linux-$ARCH_L.sh" "SHA256")
-	use_checksum_comparison "cmakeinstall.sh" "$external_hash"
+        # Supply chain security - CMake
+        hash_url=$(resolve_github_release_asset_url "Kitware/CMake" "endswith(\"SHA-256.txt\")" "latest")
+        external_hash=$(get_checksum_from_url "$hash_url" "linux-$ARCH_L.sh" "SHA256")
+        use_checksum_comparison "cmakeinstall.sh" "$external_hash"
 
-	# Install CMake and remove the install script
-	chmod +x cmakeinstall.sh \
-	&& ./cmakeinstall.sh --prefix=/usr/local --exclude-subdir \
-	&& rm cmakeinstall.sh
+        # Install CMake and remove the install script
+        chmod +x cmakeinstall.sh \
+        && ./cmakeinstall.sh --prefix=/usr/local --exclude-subdir \
+        && rm cmakeinstall.sh
+    fi
 fi
 
 invoke_tests "Tools" "Cmake"
